Fix nav issue
Observation: The nav links ul hieght was increased due to addition of new link in it. Due to which the top value of `-600%` couldn't make it hidden in collapsed state.

Fix:
In order make this dynamic, no matter how many links may be added or removed we should have the top value dynamic.

- Observation resulted in conclusion that top should be 100% multiplied by (number of links in nav - 1). So added that as inline style when the nav is in collapsed state.
- Removed top value from styles file. As that is redundant and unnecessary now.