fixed learn section issue
In learn section the subjects laravel and react are being rendered on the site itself instead of redirecting it to some other resource.

The components resposnible to display all the learn page related stuff is not getting the router object through url member in props. This is because we are using custome App.js file for entry point.

Now I'm passing the router object as url prop to child components in App so that they can get access to it and behave in the expected manner.